order export/import/summary

+++ app/Plugins/Orders/OrderController.php
@@ -7,6 +7,7 @@ use App\Plugins\Admin\AdminController;
 use App\Plugins\Orders\Functions\Orders;
 use App\Plugins\Orders\Model\OrderHeader;
 use App\Plugins\Orders\Model\OrderLines;
+use App\Plugins\Orders\Model\OriginalOrder;
 use App\Schedules;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\Schema;
@@ -27,12 +28,7 @@ class OrderController extends AdminController
             return redirect()->route($cr[0]);
         }
 
-        $trashed = false;
-        if ($cr[0] == 'orderHistory') {
-            $trashed = true;
-        }
-
-        $orders = $this->getFilteredResult($search, $trashed);
+        $orders = $this->getFilteredResult($search, $trashed ?? false);
 
         return view('admin.elements.table',
             [
@@ -42,11 +38,16 @@ class OrderController extends AdminController
                 'idField'        => 'name',
                 'destroyName'    => 'Order',
                 'filters'        => method_exists($this, 'getFilters') ? $this->getFilters() : [],
-                'logButton'    => view('admin.partials.log', ['logTypes' => 'orderExport,orderImport,createPDF,sendOrderEmails'])->render(),
+                'logButton'      => view('admin.partials.log', ['logTypes' => 'orderExport,orderImport,createPDF,sendOrderEmails,orderSummary'])->render(),
                 'currentFilters' => session('order_filters'),
                 'operations'     => view("Orders::partials.extraButtons")->render(),
                 'js'             => [
                     'js/orderlist.js',
+                ],
+                'massActions'    => [
+                    (object)['url' => route('orders.summary'), 'class' => 'btn-info', 'icon' => 'fas fa-book', 'label' => 'Order Sumary'],
+                    (object)['url' => route('orders.exportOrders'), 'class' => 'btn-warning', 'icon' => 'fas fa-download', 'label' => 'Export Orders'],
+                    (object)['url' => route('orders.destroy'), 'class' => 'btn-danger', 'callback' => 'replaceTable', 'icon' => 'fas fa-trash', 'label' => 'Destroy'],
                 ],
             ]);
     }
@@ -54,7 +55,7 @@ class OrderController extends AdminController
     public function setPaid($id)
     {
         request()->validate(['amount' => 'numeric']);
-        /** @var OrderHeader $o */
+        /** @var OrderHeader $oh */
         $oh = OrderHeader::find($id);
         $oh->update(['paid' => request('amount')]);
 
@@ -76,19 +77,18 @@ class OrderController extends AdminController
 
         $id = $id ? [$id] : request('massAction');
 
+        if (!count($id)) {
+            return ['stsatus' => false, 'message' => 'No Order specified'];
+        }
         $forceDelete = OrderHeader::onlyTrashed()->whereIn('id', $id)->forceDelete();
         $delete = OrderHeader::whereIn('id', $id)->delete();
 
         if ($forceDelete) {
             OrderLines::whereIn('order_header_id', $id)->delete();
+            OriginalOrder::find($id)->delete();
             $result = $forceDelete;
         } else {
             $result = $delete;
-        }
-
-        $trashed = false;
-        if (Route::currentRouteName() == 'orderHistory') {
-            $trashed = true;
         }
 
         /** @var array $repTable */
@@ -96,15 +96,37 @@ class OrderController extends AdminController
             [
                 'tableHeaders' => $this->getList(),
                 'header'       => 'Orders',
-                'list'         => $this->getFilteredResult(request()->get('search'), $trashed??false)->withPath(route($route)),
+                'list'         => $this->getFilteredResult(request()->get('search'), $trashed ?? false)->withPath(route($route)),
                 'idField'      => 'name',
                 'destroyName'  => 'Order',
                 'js'           => [
                     'js/orderlist.js',
                 ],
+                'massActions'  => [
+                    (object)['url' => route('orders.summary'), 'class' => 'btn-info', 'icon' => 'fas fa-book', 'label' => 'Order Sumary'],
+                    (object)['url' => route('orders.exportOrders'), 'class' => 'btn-warning', 'icon' => 'fas fa-download', 'label' => 'Export Orders'],
+                    (object)['url' => route('orders.destroy'), 'class' => 'btn-danger', 'callback' => 'replaceTable', 'icon' => 'fas fa-trash', 'label' => 'Destroy'],
+                ],
             ])->renderSections();
 
-        return ['status' => $result??true, 'message' => 'Order Deleted', 'replaceTable' => $repTable['content']];
+        return ['status' => $result ?? true, 'message' => 'Order Deleted', 'replaceTable' => $repTable['content']];
+    }
+
+
+
+    public function updateField($id, $original)
+    {
+
+        request()->validate([
+            'update' => 'required',
+        ]);
+
+        $field = request()->get('update');
+        OrderHeader::findOrFail($id)->update([$field => request()->get($field)]);
+
+        $names = $this->getViewFields();
+
+        return ['status' => true, 'message' => $names[$field] . " Updated"];
     }
 
     public function removeOrder($id = false)
@@ -112,26 +134,25 @@ class OrderController extends AdminController
         $this->destroy($id, 'forceDelete', 'orderHistory');
     }
 
-    public function showOriginal($order_id)
+    public function showOrder($order_id, $original = false)
     {
-        /** @var OrderHeader $oh */
-        $oh = new OrderHeader();
-        /** @var OrderLines $oi */
-        $oi = new OrderLines();
+        if ($original == 'original') {
+            $orderData = OriginalOrder::find($order_id);
+            /** @var OrderHeader $orderHeaders */
+            $orderHeaders = new OrderHeader($orderData->headers);
+            foreach ($orderData->items as $item) {
+                $orderItems[] = $i = new OrderLines($item);
+            }
 
-        $originalHeaders = $oh->setTable('original_headers')->findOrFail($order_id);
-        $originalItems = $oi->setTable('original_lines')->where('order_header_id', $originalHeaders->id)->get();
+        } else {
+            /** @var OrderHeader $orderHeaders */
+            $orderHeaders = OrderHeader::findOrFail($order_id);
+            $orderItems = $orderHeaders->items()->get();
+        }
 
-        return view('Orders::viewOrder', ['order' => $originalHeaders, 'items' => $originalItems, 'fields' => $this->getViewFields()]);
-    }
+        $orderHeaders->isOriginal = $original;
 
-    public function showOrder($order_id)
-    {
-        /** @var OrderHeader $orderHeaders */
-        $orderHeaders = OrderHeader::findOrFail($order_id);
-        $orderItems = $orderHeaders->items()->get();
-
-        return view('Orders::viewOrder', ['order' => $orderHeaders, 'items' => $orderItems, 'fields' => $this->getViewFields(), 'canEdit' => "editable"]);
+        return view('Orders::viewOrder', ['order' => $orderHeaders, 'items' => $orderItems, 'fields' => $this->getViewFields(), 'canEdit' => "editable", 'original' => $original]);
     }
 
     public function changeDelivery($id)
@@ -209,10 +230,15 @@ class OrderController extends AdminController
 
     public function exportOrders()
     {
-        if (!session('order_filters.market_day')) {
-            return ['status' => false, 'message' => 'Please select market Day'];
+        if(request()->has('massAction')) {
+            $orders = OrderHeader::findMany(request('massAction'))->pluck('id')->toArray();
+        } else {
+            if (!session('order_filters.market_day')) {
+                return ['status' => false, 'message' => 'Please select market Day'];
+            }
+
+            $orders = $this->getFilteredResult(request()->route('search'), false, true)->pluck('id')->toArray();
         }
-        $orders = $this->getFilteredResult(request()->route('search'), false, true)->pluck('id')->toArray();
 
         Schedules::create([
             'filename' => json_encode($orders),
@@ -220,6 +246,26 @@ class OrderController extends AdminController
         ]);
 
         return ['status' => true, 'message' => 'Export Scheduled, please consult import/export log for a download link in few minutes'];
+    }
+
+    public function createSummary() {
+        if(request()->has('massAction')) {
+            $orders = OrderHeader::findMany(request('massAction'))->pluck('id')->toArray();
+        } else {
+            if (!session('order_filters.market_day')) {
+                return ['status' => false, 'message' => 'Please select market Day'];
+            }
+
+            $orders = $this->getFilteredResult(request()->route('search'), false, true)->pluck('id')->toArray();
+        }
+
+        Schedules::create([
+            'filename' => json_encode($orders),
+            'type'     => 'orderSummary',
+        ]);
+
+        return ['status' => true, 'message' => 'Summary Scheduled, please consult import/export log for a download link in few minutes'];
+
     }
 
     public function importOrders()
